requirements.txt
google-cloud-error-reporting==0.33.0
beautifulsoup4==4.9.1
fake-useragent==0.1.11
requests==2.23.0
responses==0.10.14
